Fix bug in login

+++ database/migrations/2025_07_15_220757_add_login_and_otp_fields_to_users_table.php
@@ -4,15 +4,11 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-    /**
-     * Run the migrations.
-     */
+return new class extends Migration {
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedTinyInteger('login_attempts')->default(0);
+            $table->unsignedInteger('login_attempts')->default(0);
             $table->timestamp('last_failed_login_at')->nullable();
             $table->unsignedTinyInteger('otp_attempts')->default(0);
             $table->boolean('is_locked')->default(false);
@@ -26,3 +22,4 @@ return new class extends Migration
         });
     }
 };
+
